refactor: restructure networking with validation and organized packets
Reorganized networking layer to support robust client-server communication:
- Created packets/clientbound and packets/serverbound directories
- Added BasePacket with messageIsValid flag for decode validation
- Implemented ClientPacketHandler and ServerPacketHandler with side validation
- Moved business logic from handlers into packet handle() methods
- Created ClientDataManager to separate data management from packet handling
- Added comprehensive error logging for decode failures and wrong-side packets
- NetworkHandler now logs packet registration count

All packets now extend BasePacket, validate on decode with try-catch blocks,
and execute business logic directly in handle() after validation passes.
Handlers are now single-responsibility validators only.
