+++ src/main/java/org/sosly/villagetale/client/ClientEventHandler.java
@@ -6,7 +6,6 @@ import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import org.sosly.villagetale.VillageTale;
-import org.sosly.villagetale.network.ClientPacketHandler;
 
 @OnlyIn(Dist.CLIENT)
 @Mod.EventBusSubscriber(modid = VillageTale.MOD_ID, value = Dist.CLIENT)
@@ -14,6 +13,6 @@ public class ClientEventHandler {
 
     @SubscribeEvent
     public static void onClientDisconnect(ClientPlayerNetworkEvent.LoggingOut event) {
-        ClientPacketHandler.clearCache();
+        ClientDataManager.clearAll();
     }
 }
