+++ src/main/java/org/sosly/villagetale/network/BasePacket.java
@@ -1,0 +1,12 @@
+package org.sosly.villagetale.network;
+
+public abstract class BasePacket {
+    protected boolean messageIsValid = false;
+
+    public BasePacket() {
+    }
+
+    public final boolean isMessageValid() {
+        return this.messageIsValid;
+    }
+}
