+++ src/main/java/org/sosly/villagetale/network/NetworkHandler.java
@@ -10,6 +10,10 @@ import org.sosly.villagetale.VillageTale;
 import org.sosly.villagetale.entity.Villager;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.item.ItemStack;
+import org.sosly.villagetale.network.packets.clientbound.VillagerEquipmentSyncPacket;
+import org.sosly.villagetale.network.packets.clientbound.VillagerProfessionSyncPacket;
+import org.sosly.villagetale.network.packets.clientbound.VillageBoundaryPacket;
+import org.sosly.villagetale.network.packets.clientbound.ZoneBoundaryPacket;
 
 public class NetworkHandler {
     private static final String PROTOCOL_VERSION = "1";
@@ -54,6 +58,8 @@ public class NetworkHandler {
             ZoneBoundaryPacket::decode,
             ZoneBoundaryPacket::handle
         );
+
+        VillageTale.LOGGER.info("VillageTale registered {} network messages", packetId);
     }
 
     public static void syncProfessionToPlayer(Villager villager, ServerPlayer player) {
