+++ src/main/java/org/sosly/villagetale/network/ServerPacketHandler.java
@@ -1,19 +1,16 @@
 package org.sosly.villagetale.network;
 
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.fml.LogicalSide;
 import net.minecraftforge.network.NetworkEvent;
 import org.slf4j.Logger;
 import org.sosly.villagetale.VillageTale;
 
-@OnlyIn(Dist.CLIENT)
-public class ClientPacketHandler {
+public class ServerPacketHandler {
     public static <T extends BasePacket> boolean validateBasics(T message, NetworkEvent.Context ctx) {
         LogicalSide sideReceived = ctx.getDirection().getReceptionSide();
         ctx.setPacketHandled(true);
 
-        if (sideReceived != LogicalSide.CLIENT) {
+        if (sideReceived != LogicalSide.SERVER) {
             Logger logger = VillageTale.LOGGER;
             String messageName = message.getClass().getName();
             logger.error(messageName + " received on wrong side: " + sideReceived);
